"chore: refactor fireTrail by actions"
